+++ src/components/UploadQuestions/index.module.scss
@@ -1,4 +1,4 @@
-@import "../../../assets/RouteStyle/base.scss";
+@import "../../assets/RouteStyle/base.scss";
 
 
 .modal__container,
