+++ src/routes/UploadQuestion.jsx
@@ -1,11 +1,12 @@
 import React from 'react'
 import Layout from '../components/Layout'
-import Modal from '../components/UploadQuestions/Modal'
+import Question from '../components/ViewQuestions/Modal'
+// import Modal from '../components/UploadQuestions/Modal'
 
 export default function UploadQuestion() {
     return (
         <Layout>
-            <Modal/>
+            <Question/>
         </Layout>
     )
 }
